Default to exit code 2 on help message
When a command fails because it's mis-used and shows the help message instead of executing normally, it should *never* fail with exit code 0 as that will improperly foul things up if the command is executed in an automated way (i.e. a script or `make`).

The convention is code 2, which means "misuse".